+++ R/mod_tabpanel_linelist.R
@@ -84,6 +84,11 @@ mod_tabpanel_linelist_server <- function(
           ),
           bslib::card(
             min_height = "500px",
+            shiny::h1("Comparisons table"),
+            DT::DTOutput(ns("comparisons_tbl"))
+          ),
+          bslib::card(
+            min_height = "500px",
             shiny::h1("Case Linelist for Selected Signals"),
             shiny::span("Export or review cases linked to the selected signals."),
             DT::DTOutput(ns("linelist"))
@@ -159,6 +164,19 @@ mod_tabpanel_linelist_server <- function(
         signals_padded = signals_padded(),
         filtered_data = filtered_data()
       )
+    })
+
+    comparisons_table <- shiny::reactive({
+      req(cases_linelist)
+
+      build_comparison_summary(cases_linelist())
+    })
+
+    output$comparisons_tbl <- DT::renderDT({
+      req(comparisons_table)
+
+      comparisons_table() %>%
+        DT::datatable(rownames = FALSE, colnames = c("Measure", "Cases in selected signals", "Rest of cases"))
     })
 
     # display line lists of selected signals