+++ NEWS.md
@@ -6,6 +6,7 @@
 * Introduced a post-processing score for alarms to support priority-based filtering
 * Fixed bug in linelist tab that handled NA stratum values wrongly for selected signals.
 * Added visualisation for Signal score in time series plots
+* Implemented Comparisons table in the linelist tab, showing summary statistics for age and sex ratios for cases in selected signals.
 
 # SignalDetectionTool 0.10.0